Multivariate computation.

Easily compute multivariate sum, mean, and other scores. Reverse scoring can also be easily implemented without saving extra variables. Alpha function uses a similar method to deal with reverse scoring.

Three options to specify variables:

var + items: use the common and unique parts of variable names.

vars: directly define a character vector of variables.

varrange: use the starting and stopping positions of variables.

COUNT(data, var = NULL, items = NULL, vars = NULL, varrange = NULL, value = NA)

MODE(data, var = NULL, items = NULL, vars = NULL, varrange = NULL)

SUM(
  data,
  var = NULL,
  items = NULL,
  vars = NULL,
  varrange = NULL,
  rev = NULL,
  likert = NULL,
  na.rm = TRUE
)

MEAN(
  data,
  var = NULL,
  items = NULL,
  vars = NULL,
  varrange = NULL,
  rev = NULL,
  likert = NULL,
  na.rm = TRUE
)

STD(
  data,
  var = NULL,
  items = NULL,
  vars = NULL,
  varrange = NULL,
  rev = NULL,
  likert = NULL,
  na.rm = TRUE
)

CONSEC(
  data,
  var = NULL,
  items = NULL,
  vars = NULL,
  varrange = NULL,
  values = 0:9
)

Arguments

data

Data frame.

var

[Option 1] The common part across the variables. e.g., "RSES"

items

[Option 1] The unique part across the variables. e.g., 1:10

vars

[Option 2] A character vector specifying the variables. e.g., c("X1", "X2", "X3", "X4", "X5")

varrange

[Option 3] A character string specifying the positions ("starting:stopping") of variables. e.g., "A1:E5"

value

[Only for COUNT] The value to be counted.

rev

[Optional] Variables that need to be reversed. It can be (1) a character vector specifying the reverse-scoring variables (recommended), or (2) a numeric vector specifying the item number of reverse-scoring variables (not recommended).

likert

[Optional] Range of likert scale (e.g., 1:5, c(1, 5)). If not provided, it will be automatically estimated from the given data (BUT you should use this carefully).

na.rm

Ignore missing values. Default is TRUE.

values

[Only for CONSEC] Values to be counted as consecutive identical values. Default is all numbers (0:9).

Value

A vector of computed values.

Functions

COUNT: Count a certain value across multiple variables.

MODE: Compute mode across multiple variables.

SUM: Compute sum across multiple variables.

MEAN: Compute mean across multiple variables.

STD: Compute standard deviation across multiple variables.

CONSEC: Compute consecutive identical digits across multiple variables (especially useful in detecting careless responding).

Examples

d=data.table(x1=1:5,
             x4=c(2,2,5,4,5),
             x3=c(3,2,NA,NA,5),
             x2=c(4,4,NA,2,5),
             x5=c(5,4,1,4,5))
d
#>    x1 x4 x3 x2 x5
#> 1:  1  2  3  4  5
#> 2:  2  2  2  4  4
#> 3:  3  5 NA NA  1
#> 4:  4  4 NA  2  4
#> 5:  5  5  5  5  5
## I deliberately set this order to show you
## the difference between "vars" and "varrange".

d[,`:=`(
  na=COUNT(d, "x", 1:5, value=NA),
  n.2=COUNT(d, "x", 1:5, value=2),
  sum=SUM(d, "x", 1:5),
  m1=MEAN(d, "x", 1:5),
  m2=MEAN(d, vars=c("x1", "x4")),
  m3=MEAN(d, varrange="x1:x2", rev="x2", likert=1:5),
  cons1=CONSEC(d, "x", 1:5),
  cons2=CONSEC(d, varrange="x1:x5")
)]
#>    x1 x4 x3 x2 x5 na n.2 sum  m1  m2 m3 cons1 cons2
#> 1:  1  2  3  4  5  0   1  15 3.0 1.5  2     0     0
#> 2:  2  2  2  4  4  0   3  14 2.8 2.0  2     2     3
#> 3:  3  5 NA NA  1  2   0   9 3.0 4.0  4     0     0
#> 4:  4  4 NA  2  4  1   1  14 3.5 4.0  4     2     2
#> 5:  5  5  5  5  5  0   0  25 5.0 5.0  4     5     5
d
#>    x1 x4 x3 x2 x5 na n.2 sum  m1  m2 m3 cons1 cons2
#> 1:  1  2  3  4  5  0   1  15 3.0 1.5  2     0     0
#> 2:  2  2  2  4  4  0   3  14 2.8 2.0  2     2     3
#> 3:  3  5 NA NA  1  2   0   9 3.0 4.0  4     0     0
#> 4:  4  4 NA  2  4  1   1  14 3.5 4.0  4     2     2
#> 5:  5  5  5  5  5  0   0  25 5.0 5.0  4     5     5

data=as.data.table(psych::bfi)
data[,`:=`(
  E=MEAN(d, "E", 1:5, rev=c(1,2), likert=1:6),
  O=MEAN(d, "O", 1:5, rev=c(2,5), likert=1:6)
)]
#>       A1 A2 A3 A4 A5 C1 C2 C3 C4 C5 E1 E2 E3 E4 E5 N1 N2 N3 N4 N5 O1 O2 O3 O4
#>    1:  2  4  3  4  4  2  3  3  4  4  3  3  3  4  4  3  4  2  2  3  3  6  3  4
#>    2:  2  4  5  2  5  5  4  4  3  4  1  1  6  4  3  3  3  3  5  5  4  2  4  3
#>    3:  5  4  5  4  4  4  5  4  2  5  2  4  4  4  5  4  5  4  2  3  4  2  5  5
#>    4:  4  4  6  5  5  4  4  3  5  5  5  3  4  4  4  2  5  2  4  1  3  3  4  3
#>    5:  2  3  3  4  5  4  4  5  3  2  2  2  5  4  5  2  3  4  4  3  3  3  4  3
#>   ---                                                                        
#> 2796:  6  1  3  3  3  6  6  6  1  1  1  4  5  5  6  1  1  1 NA  1  6  1  6  6
#> 2797:  2  4  4  3  5  2  3  4  4  3  2  2  4  4  3 NA  3  2  3  3  6  3  5  4
#> 2798:  2  3  5  2  5  5  5  5  1  1  2  2  6  3  6  3  4  3  3  1  5  1  6  4
#> 2799:  5  2  2  4  4  5  5  5  2  6  2  2  4  5  4  5  5  6  4  1  5  2  5  5
#> 2800:  2  3  1  4  2  5  5  3  3  3  3  3  1  2  2  1  2  2  1  1  3  1  3  5
#>       O5 gender education age E O
#>    1:  3      1        NA  16    
#>    2:  3      2        NA  18    
#>    3:  2      2        NA  17    
#>    4:  5      2        NA  17    
#>    5:  3      1        NA  17    
#>   ---                            
#> 2796:  1      1         3  19    
#> 2797:  2      1         4  27    
#> 2798:  3      2         4  29    
#> 2799:  1      1         4  31    
#> 2800:  1      2         4  50    
data
#>       A1 A2 A3 A4 A5 C1 C2 C3 C4 C5 E1 E2 E3 E4 E5 N1 N2 N3 N4 N5 O1 O2 O3 O4
#>    1:  2  4  3  4  4  2  3  3  4  4  3  3  3  4  4  3  4  2  2  3  3  6  3  4
#>    2:  2  4  5  2  5  5  4  4  3  4  1  1  6  4  3  3  3  3  5  5  4  2  4  3
#>    3:  5  4  5  4  4  4  5  4  2  5  2  4  4  4  5  4  5  4  2  3  4  2  5  5
#>    4:  4  4  6  5  5  4  4  3  5  5  5  3  4  4  4  2  5  2  4  1  3  3  4  3
#>    5:  2  3  3  4  5  4  4  5  3  2  2  2  5  4  5  2  3  4  4  3  3  3  4  3
#>   ---                                                                        
#> 2796:  6  1  3  3  3  6  6  6  1  1  1  4  5  5  6  1  1  1 NA  1  6  1  6  6
#> 2797:  2  4  4  3  5  2  3  4  4  3  2  2  4  4  3 NA  3  2  3  3  6  3  5  4
#> 2798:  2  3  5  2  5  5  5  5  1  1  2  2  6  3  6  3  4  3  3  1  5  1  6  4
#> 2799:  5  2  2  4  4  5  5  5  2  6  2  2  4  5  4  5  5  6  4  1  5  2  5  5
#> 2800:  2  3  1  4  2  5  5  3  3  3  3  3  1  2  2  1  2  2  1  1  3  1  3  5
#>       O5 gender education age E O
#>    1:  3      1        NA  16    
#>    2:  3      2        NA  18    
#>    3:  2      2        NA  17    
#>    4:  5      2        NA  17    
#>    5:  3      1        NA  17    
#>   ---                            
#> 2796:  1      1         3  19    
#> 2797:  2      1         4  27    
#> 2798:  3      2         4  29    
#> 2799:  1      1         4  31    
#> 2800:  1      2         4  50